requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
plaid-python
python-dotenv
requests
pandas
matplotlib
prophet
pydantic